refactor(infer): promote CST declaration/receiver helpers out of semantic_tokens
Shared home for classify_symbol_at (slice 6) — avoids a second CST pass
answering the identical declaration-vs-reference and receiver-extraction
questions semantic_tokens already solved.
