+++ src/indexer/infer/cst_symbol.rs
@@ -1,0 +1,64 @@
+//! Shared CST identifier classification: declaration-vs-reference, and
+//! receiver/member extraction from a `navigation_expression`.
+//!
+//! Originally written for semantic-token coloring (`semantic_tokens/resolve.rs`);
+//! promoted here because `classify_symbol_at` (the navigation-feature
+//! classifier: go-def, goto-impl, highlight) needs the identical walk —
+//! two independent CST passes answering "declaration or reference?" and
+//! "what's the receiver of this member access?" would drift from each other.
+
+use tree_sitter::Node;
+
+use crate::queries::{
+    KIND_CLASS_DECL, KIND_CLASS_PARAM, KIND_COMPANION_OBJ, KIND_ENUM_ENTRY, KIND_FUN_DECL,
+    KIND_OBJECT_DECL, KIND_PARAMETER, KIND_SIMPLE_IDENT, KIND_TYPE_ALIAS, KIND_TYPE_IDENT,
+    KIND_TYPE_PARAM, KIND_VAR_DECL,
+};
+
+pub(crate) fn is_declaration_site(node: Node<'_>) -> bool {
+    let Some(parent) = node.parent() else {
+        return false;
+    };
+    let pk = parent.kind();
+    if pk == KIND_CLASS_DECL
+        || pk == KIND_OBJECT_DECL
+        || pk == KIND_COMPANION_OBJ
+        || pk == KIND_TYPE_ALIAS
+    {
+        return node.kind() == KIND_TYPE_IDENT;
+    }
+    if pk == KIND_FUN_DECL
+        || pk == KIND_PARAMETER
+        || pk == KIND_ENUM_ENTRY
+        || pk == KIND_VAR_DECL
+        || pk == KIND_CLASS_PARAM
+    {
+        return node.kind() == KIND_SIMPLE_IDENT;
+    }
+    if pk == KIND_TYPE_PARAM {
+        return node.kind() == KIND_SIMPLE_IDENT || node.kind() == KIND_TYPE_IDENT;
+    }
+    false
+}
+
+pub(crate) fn navigation_receiver_node(node: Node<'_>) -> Option<Node<'_>> {
+    (0..node.child_count())
+        .filter_map(|i| node.child(i))
+        .find(|child| child.is_named() && child.kind() != crate::queries::KIND_NAV_SUFFIX)
+}
+
+pub(crate) fn navigation_member_ident(node: Node<'_>) -> Option<Node<'_>> {
+    use crate::indexer::NodeExt;
+    let suffix = node.first_child_of_kind(crate::queries::KIND_NAV_SUFFIX)?;
+    (0..suffix.child_count())
+        .filter_map(|i| suffix.child(i))
+        .find(|child| child.kind() == KIND_SIMPLE_IDENT || child.kind() == KIND_TYPE_IDENT)
+}
+
+pub(crate) fn is_call_callee(node: Node<'_>) -> bool {
+    let Some(parent) = node.parent() else {
+        return false;
+    };
+    parent.kind() == crate::queries::KIND_CALL_EXPR
+        && parent.child(0).map(|child| child.id()) == Some(node.id())
+}
