+++ src/semantic_tokens/resolve.rs
@@ -16,11 +16,14 @@ use crate::resolver::infer::find_field_type_in_class;
 use crate::Language;
 
 use super::helpers::{
-    is_annotation_reference, is_call_callee, is_declaration_site, is_inside_lambda_parameters,
-    is_named_argument_label, is_navigation_receiver, is_top_level_call_name, is_type_reference,
-    navigation_member_ident, navigation_receiver_node, node_text, push_token, visit_tree,
+    is_annotation_reference, is_inside_lambda_parameters, is_named_argument_label,
+    is_navigation_receiver, is_top_level_call_name, is_type_reference, node_text, push_token,
+    visit_tree,
 };
 use super::{modifier_bit, type_index, RawToken, Source};
+use crate::indexer::{
+    is_call_callee, is_declaration_site, navigation_member_ident, navigation_receiver_node,
+};
 
 /// Walk non-declaration identifiers and resolve them against the index.
 pub(super) fn walk_references(
